Split readablestream out into its own module
